support provisioning for external cluster controller
