+++ pkg/controller/master-controller-manager/external-cluster/cluster_controller.go
@@ -26,8 +26,10 @@ import (
 
 	providerconfig "github.com/kubermatic/machine-controller/pkg/providerconfig/types"
 	kubermaticapiv1 "k8c.io/kubermatic/v2/pkg/api/v1"
+	apiv2 "k8c.io/kubermatic/v2/pkg/api/v2"
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/crd/kubermatic/v1"
 	kuberneteshelper "k8c.io/kubermatic/v2/pkg/kubernetes"
+	"k8c.io/kubermatic/v2/pkg/provider"
 	"k8c.io/kubermatic/v2/pkg/provider/cloud/aks"
 	"k8c.io/kubermatic/v2/pkg/provider/cloud/eks"
 	"k8c.io/kubermatic/v2/pkg/provider/cloud/gke"
@@ -116,12 +118,23 @@ func (r *Reconciler) Reconcile(ctx context.Context, request reconcile.Request) (
 
 func (r *Reconciler) reconcile(ctx context.Context, cluster *kubermaticv1.ExternalCluster) (reconcile.Result, error) {
 	cloud := cluster.Spec.CloudSpec
+	secretKeySelector := provider.SecretKeySelectorValueFuncFactory(ctx, r.Client)
 	if cloud == nil {
 		return reconcile.Result{}, nil
 	}
 	if cloud.GKE != nil {
 		r.log.Debugf("reconcile GKE cluster")
-		err := r.createOrUpdateGKEKubeconfig(ctx, cluster)
+		status, err := gke.GetGKEClusterStatus(ctx, secretKeySelector, cloud)
+		if err != nil {
+			r.log.Errorf("failed to get GKE cluster status %v", err)
+			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
+			return reconcile.Result{}, err
+		}
+		if status.State == apiv2.PROVISIONING {
+			// repeat after some time to get/store kubeconfig
+			return reconcile.Result{RequeueAfter: time.Second * 10}, err
+		}
+		err = r.createOrUpdateGKEKubeconfig(ctx, cluster)
 		if err != nil {
 			r.log.Errorf("failed to create or update kubeconfig secret %v", err)
 			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
@@ -132,7 +145,18 @@ func (r *Reconciler) reconcile(ctx context.Context, cluster *kubermaticv1.Extern
 	}
 	if cloud.EKS != nil {
 		r.log.Debugf("reconcile EKS cluster %v", cluster.Name)
-		err := r.createOrUpdateEKSKubeconfig(ctx, cluster)
+		status, err := eks.GetEKSClusterStatus(secretKeySelector, cloud)
+		if err != nil {
+			r.log.Errorf("failed to get EKS cluster status %v", err)
+			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
+			return reconcile.Result{}, err
+		}
+		if status.State == apiv2.PROVISIONING {
+			// repeat after some time to get/store kubeconfig
+			return reconcile.Result{RequeueAfter: time.Second * 10}, err
+		}
+
+		err = r.createOrUpdateEKSKubeconfig(ctx, cluster)
 		if err != nil {
 			r.log.Errorf("failed to create or update kubeconfig secret %v", err)
 			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
@@ -143,7 +167,17 @@ func (r *Reconciler) reconcile(ctx context.Context, cluster *kubermaticv1.Extern
 	}
 	if cloud.AKS != nil {
 		r.log.Debugf("reconcile AKS cluster %v", cluster.Name)
-		err := r.createOrUpdateAKSKubeconfig(ctx, cluster)
+		status, err := aks.GetAKSClusterStatus(ctx, secretKeySelector, cloud)
+		if err != nil {
+			r.log.Errorf("failed to get AKS cluster status %v", err)
+			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
+			return reconcile.Result{}, err
+		}
+		if status.State == apiv2.PROVISIONING {
+			// repeat after some time to get/store kubeconfig
+			return reconcile.Result{RequeueAfter: time.Second * 10}, err
+		}
+		err = r.createOrUpdateAKSKubeconfig(ctx, cluster)
 		if err != nil {
 			r.log.Errorf("failed to create or update kubeconfig secret %v", err)
 			r.recorder.Event(cluster, corev1.EventTypeWarning, "ReconcilingError", err.Error())
